Added the SDL Library to the project.  This includes adding the proper .h files to includes/sdl and adding library files to lib/windows and dlls to lib/windows/dlls
The library and dll files are only really needed for me because I am developing primarily on windows, and it is much easier to have a common repository for all the needed files since cmake throws a fit when trying to find the needed files in windows.  CMake is fine when finding the files on linux if you install them to the default directory though.