Added a nifty little toString() method like java, C++ gods don't strike me down :P
Added an ImageLoader that will handle all image loading/resizing.  This holds the actual image data.  When an image is drawn to screen, the drawer called ImageLoader::getImage(string src) which returns the image data to draw.  Src is returned whenever the image is loaded.  This is an easy way to make sure there is only one copy of an image in memory at any one time.  The only problem is once an image is loaded, it stays in memory till the game terminates.  Going to keep it like that for now since fixing that little issue is not worth the time for the limited amount of images this game uses.