Started working on GUI components for the game.
Component - the base of all gui components, abstract class that acts more like a contract for all future gui components.

UI - this holds other gui components.  Do not know if this is a moot point if I have states anyway, but it might be nice to only give events to and draw a single object in the state for all ui elements, especially if there are any HUD's in the main simulation